requirements.txt
oauth2client==3.0.0
GitPython==3.1.24
folium==0.2.1
urllib3
selenium
random-user-agent
webdriver_manager
webdriver-setup
pyvirtualdisplay
pandas
sqlalchemy
apache-airflow[amazon]
pyarrow
fastparquet
numpy
crypto==1.4.1
pycryptodome
sai-airflow-plugins
python-dateutil
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
apache-airflow-providers-google
df2gspread
s3fs
airflow-dbt